+++ src/routes/transactions.tsx
@@ -1,17 +1,15 @@
 // React
-import {useOutletContext} from "react-router-dom";
+import {useAppContext} from "@/context.ts";
 // Types
 import {Travel} from "@/types.ts";
-// Routes
-import {RootContextType} from "@/routes/root.tsx";
 // Components
 import TransactionList from "@/components/Transaction/TransactionList.tsx";
 import TransactionListSkeleton from "@/components/Transaction/TransactionListSkeleton.tsx";
 
 
 function Transactions() {
-    // Retrieve currentTravel and its update function
-    const {currentTravel} = useOutletContext<RootContextType>();
+    // Retrieve currentTravel
+    const {currentTravel} = useAppContext();
 
     function displayTransactions(travel: Travel | undefined) {
         if (travel) {