+++ CHANGELOG.md
@@ -6,13 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.5.0] - ???
+### Added
+- Create a custom hook `useAppContext` to handle the application state
+
 ### Changed
 - Updated migrations comments
+- Removed the use of `useOutletContext` in favour of `useContext`
 
 ## [0.4.0] - 2024-06-09 
 ### Added
 - User can now add and edit transactions made in a travel
-
 
 - Added React Router
 - Added pages to `Dashboard` and `Transactions`